char2int.fast {NEArender}	R Documentation

char2int function

Description

This function reads input list objects and maps unique numbers to each member in the network/ genesets.

Usage

char2int.fast(net.list, gs.list.1, gs.list.2 = NULL, Parallelize = 1)

Arguments

net.list
the global network object, pre-created using import.net

gs.list.1
AGS or FGS object that lists members of each individual AGS/FGS, pre-created using samples2ags

gs.list.2
FGS or AGS object that lists members of each individual FGS/AGS, pre-created using samples2ags

Parallelize
The number of CPU cores to be used for the step "Counting actual network links" (the other steps are sufficiently fast). The option is not supported on Windows.

See Also

nea.render

[Package NEArender version 1.5 Index]