connectivity {NEArender}	R Documentation

Connectivity

Description

Function for plotting node degree distribution in order to test if the network is scale-free

Usage

connectivity(NET, Lowercase = 1, col.1 = 1, col.2 = 2, col.score = 3,
  echo = 1, main = "Connectivity plot", min.score = NA, n.top = NA,
  hex = FALSE)

Arguments

NET
Input network file.

Lowercase
If node IDs should be rendered lower-case (Default:1, i.e. 'yes').

col.1
Number of column where 1st node of each edge should be found (only needed when NET is a text file rather than a list).

col.2
Number of column where 2nd node of each edge should be found (only needed when NET is a text file rather than a list, i.e. passed down to import.net).

col.score
Number of column where edge confidence score is found (only needed when NET is a text file rather than a list, i.e. passed down to import.net).

echo
If messages about execution progress should appear.

main
title name for the plot, default: "Connectivity plot"

min.score
Minimum confidence score for an edge to be included in the network (is alternative to n.top and is only used when NET is a text file rather than a list, i.e. passed down to import.net).

n.top
Number of edges to be included in the network, top when ranked by confidence score (is alternative to min.score and only used when NET is a text file rather than a list, i.e. passed down to import.net).

hex
If the node degree distribution should be presented as density plot using package hexbin.

See Also

benchmark and import.net

Examples

file <-  system.file("extdata", "Connectivity.FC1_full", package = "NEArender")

connect <- connectivity(file,hex=TRUE)

[Package NEArender version 1.5 Index]