import.gs {NEArender}	R Documentation

Read in an AGS or FGS test file

Description

Imports a TAB-delimited AGS or FGS file. Checks the format of the given FGS file. Reads in unique gene/protein IDs from the given file tbl so that the created object can be directly submitted to nea.render and gsea.render. The function import.gs will be automatically called from nea.render if parameters AGS and/or FGS of the latter correspond to text files rather than lists.

Usage

import.gs(tbl, Lowercase = 1, col.gene = 2, col.set = 3, gs.type = "",
  header = FALSE)

Arguments

tbl
Input TAB-delimited FGS file.

Lowercase
render gene/protein IDs lower-case.

col.gene
Number of the column with gene/protein IDs in the input file.

col.set
Number of the column with set (sample, pathway etc.) IDs in the input file. Special options: col.set = 0: all gene/protein IDs in the file become members of a single FGS; col.set < 0 each single gene/protein IDs becomes a separate FGS (ID of which equals the gene/protein ID).

gs.type
GS type, 'a' produces AGS, 'f' produces FGS.

header
assign logical value (TRUE/FALSE) to consider the first line as header from the input file (default:FALSE).

Value

A list with entry names that correspond AGS/FGS IDs in the input file and gene/protein IDs as elements of each entry.

See Also

mutations2ags, samples2ags

Examples

data(can.sig.go)
fpath <- can.sig.go
fgs.list <- import.gs(fpath)
summary(fgs.list)

[Package NEArender version 1.5 Index]