import.net {NEArender}	R Documentation

Import a network text file

Description

This function reads in a TAB-delimited file with the global network data needed for NEA. Checks the format of the given NETWORK file, processes unique genes from the given network file, and counts the network links.

Usage

import.net(tbl, Lowercase = 1, col.1 = 1, col.2 = 2, col.score = 3,
  min.score = NA, n.top = NA, echo = 1)

Arguments

tbl
Input network file.

Lowercase
Render gene/protein IDs lower-case.

col.1
Number of the column in the input file that corresponds to node 1 (gene/protein ID).

col.2
Number of the column in the input file that corresponds to node 2 (gene/protein ID).

col.score
Number of column where edge confidence score is found.

min.score
Minimum confidence score for an edge to be included in the network.

n.top
Number of edges to be included in the network, top when ranked by confidence score (is alternative to min.score).

echo
if execution progress should be reported.

Value

$Ntotal, the total number of edges in the network and $links, an object with members (named by gene/protein IDs) that correspond to each node in the global network. Each entry of the list $links contains a vector of neighbours of the respective node.

See Also

mutations2ags, samples2ags

Examples

data(net.kegg)
net <- import.net(net.kegg)
summary(net$links)
print(paste(names(net$links)[100], net$links[[100]], sep=": "))

[Package NEArender version 1.5 Index]