net.kegg {NEArender}	R Documentation

A more compact alternative global network for the network enrichment analysis

Description

A more compact alternative global network for the network enrichment analysis

Usage

data(net.kegg)

Format

A list, see import.net

Examples

head(net.kegg)

[Package NEArender version 1.5 Index]