set.heat {NEArender}	R Documentation

Plot a heatmap of NEA/GSEA output

Description

Plots a heatmap where width and height of each element reflect respective sizes of input FGS and AGS.

Usage

set.heat(List1, List2, Z, Log = TRUE)

Arguments

List1
AGS or FGS object that lists members of each individual AGS/FGS.

List2
FGS or AGS object that lists members of each individual FGS/AGS.

Z
matrix (output of nea.render with Z-scores that will define coloring of the heatmap.

Log
If TRUE, then the Z values will be log-transformed (the default).

See Also

nea.render

Examples

ags.list <- samples2ags(fantom5.43samples, Ntop=20, method="topnorm")
data(can.sig.go)
fpath <- can.sig.go
fgs.list <- import.gs(fpath)
data(net.kegg)
netpath <- net.kegg
net <- import.net(netpath)
n1 <- nea.render(AGS=ags.list, FGS=fgs.list, NET=net, graph=TRUE)
set.heat(ags.list, fgs.list, n1$z, Log=FALSE)

[Package NEArender version 1.5 Index]