topology2nd {NEArender}	R Documentation

Higher order topology and correlation between node degrees

Description

Auxiliary function to estimate how non-random are associations of node degrees in network edges.

Usage

topology2nd(NET, Lowercase = 1, col.1 = 1, col.2 = 2, echo = 1,
  main = "Higher order topology")

Arguments

NET
Either a text network file or an R list imported with import.net

Lowercase
Render gene/protein IDs lower-case.

col.1
Number of the column in the input file that corresponds to node 1 (gene/protein ID).

col.2
Number of the column in the input file that corresponds to node 2 (gene/protein ID).

echo
if execution progress should be reported.

main
The plot title.

See Also

connectivity, benchmark

Examples

file <-  system.file("extdata", "Connectivity.FC1_full", package = "NEArender")

 topology2nd(NET=file)

[Package NEArender version 1.5 Index]